move components to the files where they belong
