+++ lib/src/presentation/components/map.dart
@@ -1,23 +1,13 @@
 // flutter imports
 import 'dart:developer';
-import 'dart:math' as math;
 import 'dart:ui';
-import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_background_service/flutter_background_service.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_map/plugin_api.dart';
-import 'package:geolocator/geolocator.dart';
-import 'package:latlong2/latlong.dart';
-import 'package:intl/intl.dart';
 
 // local imports
 import 'package:geo_steps/src/application/location.dart';
-import 'package:geo_steps/src/application/preferences.dart';
-import 'package:geo_steps/src/presentation/components/icons.dart';
-import 'package:geo_steps/src/presentation/components/lines.dart';
-import 'package:geo_steps/src/presentation/components/overview_stats.dart';
 import 'package:geo_steps/src/utils/map.dart';
 import 'package:geo_steps/src/utils/sizing.dart';
 
@@ -102,399 +92,7 @@ class _MapPreviewState extends State<MapPreview> {
   }
 }
 
-class TodaysMap extends StatefulWidget {
-  late DateTime date;
 
-  TodaysMap({super.key, showDate}) {
-    if (showDate != null) {
-      date = showDate;
-    } else {
-      date = DateTime.now();
-    }
-  }
-
-  @override
-  State<StatefulWidget> createState() => _TodaysMapState();
-}
-
-class _TodaysMapState extends State<TodaysMap>
-    with SingleTickerProviderStateMixin {
-  late Animation<double> detailsAnimation;
-  late AnimationController detailsController;
-
-  late LocationService locationService;
-  late TargetPlatform defaultTargetPlatform = TargetPlatform.iOS;
-  final mapController = MapController();
-  bool showDetails = false;
-  static const double mapHeightDetails = 150;
-
-  @override
-  void initState() {
-    super.initState();
-
-    locationService = LocationService();
-    locationService
-        .init()
-        .whenComplete(() => locationService.loadToday().then((wasLoaded) {
-              if (wasLoaded && locationService.hasPositions) {
-                setState(() => mapController.move(
-                    LatLng(locationService.lastPos!.latitude,
-                        locationService.lastPos!.longitude),
-                    12.8));
-              }
-            }));
-
-    AppSettings().trackingLocation.get().then((isTrackingLocation) {
-      if (isTrackingLocation != null && isTrackingLocation) {
-        FlutterBackgroundService().on("sendTrackingData").listen((event) {
-          setState(() {
-            List<dynamic> receivedDataPoints = event!["trackingData"];
-            var changed = locationService.dataPointsFromKV(receivedDataPoints);
-            if (changed) {
-              mapController.move(
-                  LatLng(locationService.lastPos!.latitude,
-                      locationService.lastPos!.longitude),
-                  14.5);
-            }
-          });
-        });
-        Timer.periodic(const Duration(seconds: 10), (timer) {
-          FlutterBackgroundService().invoke("requestTrackingData");
-        });
-      }
-    });
-
-    // locationService.record(onReady: (p) {
-    //   mapController.move(
-    //       LatLng(p.latitude, p.longitude), 12.8);
-    // });
-
-    detailsController = AnimationController(
-        vsync: this, duration: const Duration(milliseconds: 150));
-    detailsAnimation = Tween<double>(
-            begin: mapHeightDetails + 58,
-            end: SizeHelper().heightWithoutNav - 200)
-        .animate(detailsController)
-      ..addListener(() {
-        setState(() {});
-      });
-  }
-
-  @override
-  void dispose() {
-    // locationService.stopRecording();
-
-    super.dispose();
-  }
-
-  void toggleDetailsOpen() {
-    if (!showDetails) {
-      detailsController.forward();
-    } else {
-      detailsController.reverse();
-    }
-    log("before $showDetails");
-    setState(() {
-      showDetails = !showDetails;
-    });
-    log("after $showDetails");
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    defaultTargetPlatform = Theme.of(context).platform;
-    SizeHelper sizer = SizeHelper();
-    return SizedBox(
-      height: sizer.heightWithoutNav,
-      child: Stack(
-        children: [
-          Positioned(
-            child: SizedBox(
-                width: sizer.width,
-                height: sizer.heightWithoutNav - detailsAnimation.value,
-                child: FlutterMap(
-                  mapController: mapController,
-                  options: MapOptions(
-                      onMapReady: () {
-                        // set initial position on map
-                      },
-                      zoom: 13.0,
-                      maxZoom: 19.0,
-                      keepAlive: true,
-                      interactiveFlags: // all interactions except rotation
-                          InteractiveFlag.all & ~InteractiveFlag.rotate),
-                  nonRotatedChildren: [
-                    CustomAttributionWidget.defaultWidget(
-                      source: '© OpenStreetMap contributors',
-                      sourceTextStyle: const TextStyle(
-                          fontSize: 12, color: Color(0xFF0078a8)),
-                      onSourceTapped: () {},
-                    ),
-                  ],
-                  children: [
-                    TileLayer(
-                      urlTemplate:
-                          "https://tile.openstreetmap.org/{z}/{x}/{y}.png",
-                      userAgentPackageName: 'dev.janeuster.geo_steps',
-                    ),
-                    // kinda cool, shit res outside us, unknown projection - /{z}/{y}/{x} does not work
-                    // TileLayer(urlTemplate: "https://basemap.nationalmap.gov/arcgis/rest/services/USGSTopo/MapServer/tile/4/5/5?blankTile=false",
-                    //   userAgentPackageName: 'dev.janeuster.geo_steps'),
-                    // ),
-                    PolylineLayer(
-                      polylineCulling: false,
-                      polylines: [
-                        Polyline(
-                          points: locationService.latLngList,
-                          color: Colors.black,
-                          strokeWidth: 8,
-                        ),
-                        Polyline(
-                          points: locationService.latLngList,
-                          color: Colors.white,
-                          strokeWidth: 4,
-                        ),
-                      ],
-                    ),
-                    if (locationService.hasPositions)
-                      MarkerLayer(
-                        markers: [
-                          Marker(
-                              width: 46,
-                              height: 46,
-                              point: LatLng(locationService.lastPos!.latitude,
-                                  locationService.lastPos!.longitude),
-                              builder: (context) => Transform.translate(
-                                    offset: const Offset(0, -23),
-                                    child: Container(
-                                        decoration: const BoxDecoration(
-                                            image: DecorationImage(
-                                                image: AssetImage(
-                                                    "assets/map_pin.png")))),
-                                  ))
-                        ],
-                      )
-                  ],
-                )),
-          ),
-          Positioned(
-            bottom: 0,
-            child: SizedBox(
-              height: detailsAnimation.value,
-              child: Column(
-                children: [
-                  GestureDetector(
-                      onTap: () => toggleDetailsOpen(),
-                      child: Container(
-                          width: sizer.width,
-                          height: 58,
-                          padding: const EdgeInsets.all(5),
-                          color: Colors.black,
-                          child: Column(
-                            children: [
-                              Text(
-                                  "show ${showDetails ? "less" : "more"} info for ${DateFormat.yMMMMEEEEd().format(widget.date)}",
-                                  style: const TextStyle(
-                                      color: Colors.white,
-                                      fontWeight: FontWeight.bold)),
-                              const Padding(
-                                  padding: EdgeInsets.only(bottom: 5)),
-                              Transform.rotate(
-                                  angle: showDetails ? 0 : 1 * pi,
-                                  child: const Icon(
-                                    Icomoon.arrow,
-                                    color: Colors.white,
-                                  )),
-                            ],
-                          ))),
-                  SizedBox(
-                    width: sizer.width,
-                    height: detailsAnimation.value - 58,
-                    child: !showDetails
-                        ? const HourlyActivity()
-                        : ListView(
-                            children: [
-                              Padding(
-                                padding:
-                                    const EdgeInsets.only(left: 10, bottom: 15),
-                                child: Row(
-                                  children: [
-                                    OverviewTotals(timeFrameString: "Today", totalSteps: 6929, totalDistance: 4200,),
-                                    Expanded(child: Container()),
-                                  ],
-                                ),
-                              ),
-                              const Padding(
-                                  padding: EdgeInsets.symmetric(
-                                      vertical: 16, horizontal: 10),
-                                  child: DottedLine(
-                                    height: 2,
-                                  )),
-                              const HourlyActivity(),
-                              const Padding(
-                                  padding: EdgeInsets.symmetric(
-                                      vertical: 16, horizontal: 10),
-                                  child: DottedLine(
-                                    height: 2,
-                                  )),
-                            ],
-                          ),
-                  )
-                ],
-              ),
-            ),
-          )
-        ],
-      ),
-    );
-  }
-}
-
-class HourlyActivity extends StatefulWidget {
-  final double hourWidth = 50;
-
-  const HourlyActivity({super.key});
-
-  @override
-  State<StatefulWidget> createState() => _HourlyActivityState();
-}
-
-class _HourlyActivityState extends State<HourlyActivity> {
-  ScrollController scrollController =
-      ScrollController(initialScrollOffset: 715);
-  bool isScrolling = false;
-  int selectedHourIndex = 0;
-
-  @override
-  void initState() {
-    WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
-      scrollController.addListener(() {
-        if (scrollController.position.pixels ==
-                scrollController.position.maxScrollExtent ||
-            scrollController.position.pixels ==
-                scrollController.position.minScrollExtent) {
-          setSelectedHour();
-        }
-      });
-      scrollController.position.isScrollingNotifier.addListener(() {
-        if (scrollController.positions.isNotEmpty) {
-          var scrollBool = scrollController.position.isScrollingNotifier.value;
-          if (scrollBool != isScrolling) {
-            setState(() {
-              isScrolling = scrollBool;
-            });
-            if (scrollBool == false) {
-              setSelectedHour();
-            }
-          }
-        }
-      });
-      setSelectedHour();
-    });
-    super.initState();
-  }
-
-  setSelectedHour() {
-    var pixels = scrollController.position.pixels + widget.hourWidth / 2;
-    int newIndex =
-        (pixels / scrollController.position.maxScrollExtent * 23).floor();
-    setState(() {
-      selectedHourIndex = newIndex;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    var sizer = SizeHelper();
-    return SizedBox(
-        width: sizer.width,
-        height: 139,
-        child: Padding(
-          padding: const EdgeInsets.symmetric(vertical: 5, horizontal: 0),
-          child: ListView.builder(
-              itemCount: 24,
-              padding: EdgeInsets.symmetric(
-                  horizontal: sizer.width / 2 - 25, vertical: 0),
-              scrollDirection: Axis.horizontal,
-              controller: scrollController,
-              itemBuilder: (BuildContext context, int index) {
-                List<double> hoursPercent = [
-                  0,
-                  0,
-                  0,
-                  0,
-                  0,
-                  .1,
-                  .4,
-                  .6,
-                  0,
-                  0,
-                  .1,
-                  0,
-                  .3,
-                  1,
-                  .5,
-                  1,
-                  .1,
-                  .3,
-                  .2,
-                  0,
-                  0,
-                  0,
-                  0,
-                  0
-                ]; // TODO: use real data
-                return Container(
-                  padding: index < 23
-                      ? const EdgeInsets.only(right: 5)
-                      : const EdgeInsets.all(0),
-                  height: 139,
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.start,
-                    children: [
-                      const Padding(padding: EdgeInsets.only(bottom: 5)),
-                      Container(
-                          color: index == selectedHourIndex
-                              ? Colors.black
-                              : Colors.white,
-                          width: widget.hourWidth,
-                          height: 3),
-                      SizedBox(
-                          height: 105,
-                          child: Column(children: [
-                            Expanded(
-                                child: Container(
-                              width: widget.hourWidth,
-                              decoration: BoxDecoration(
-                                  border: Border.all(
-                                      color: Colors.black, width: 1)),
-                            )),
-                            Container(
-                                width: widget.hourWidth,
-                                height: 105 * hoursPercent[index],
-                                color: index == selectedHourIndex
-                                    ? null
-                                    : Colors.black,
-                                decoration: index == selectedHourIndex
-                                    ? BoxDecoration(
-                                        border: Border.all(
-                                            color: Colors.black, width: 1),
-                                        image: const DecorationImage(
-                                            fit: BoxFit.none,
-                                            scale: 2.5,
-                                            image: AssetImage(
-                                                "assets/line_pattern.jpg"),
-                                            repeat: ImageRepeat.repeat))
-                                    : null),
-                          ])),
-                      SizedBox(height: 16, child: Text("$index"))
-                    ],
-                  ),
-                );
-              }),
-        ));
-  }
-}
 
 class CustomAttributionWidget extends AttributionWidget {
   const CustomAttributionWidget({super.key, required super.attributionBuilder});
